Removing FaceVectorPD
