+++ util/mechanic/aerial_turn/face_vector.py
@@ -1,9 +1,10 @@
 from util.mechanic.base_mechanic import BaseMechanic
 from rlbot.agents.base_agent import SimpleControllerState
 from RLUtilities.Simulation import Car
-from RLUtilities.LinearAlgebra import vec3, dot, transpose, rotation_to_axis, norm
+from RLUtilities.LinearAlgebra import vec3, dot
 from RLUtilities.Maneuvers import look_at, AerialTurn
-from .PD import *
+from util.coordinates import spherical
+import math
 
 
 class BaseFaceVector(BaseMechanic):
@@ -29,6 +30,7 @@ class FaceVectorRLU(BaseFaceVector):
 
     def step(self, car: Car, target: vec3, dt: float = 0.01667) -> SimpleControllerState:
         super(FaceVectorRLU, self).step(car, target)
+        # up = vec3(*[self.car.theta[i, 2] for i in range(3)])
 
         target_rotation = look_at(self.target)
         self.action = AerialTurn(car, target_rotation)
@@ -39,23 +41,5 @@ class FaceVectorRLU(BaseFaceVector):
         self.car.last_input.pitch = self.controls.pitch
         self.car.last_input.yaw = self.controls.yaw
         self.car.last_input.roll = self.controls.roll
-
-        return self.controls
-
-
-class FaceVectorPD(BaseFaceVector):
-
-    def step(self, car: Car, target: vec3) -> SimpleControllerState:
-        super(FaceVectorPD, self).step(car, target)
-
-        local_pos = dot(self.target, self.car.theta)
-        local_pos_spherical = spherical(local_pos)
-        roll = rotationMatrixToEulerAngles(self.car.theta)[0]
-
-        local_omega = dot(self.car.omega, self.car.theta)
-
-        self.controls.pitch = pitch_point(local_pos_spherical[1], local_omega[1])
-        self.controls.yaw = yaw_point(local_pos_spherical[2], local_omega[2])
-        self.controls.roll = roll_point(roll, local_omega[0])
 
         return self.controls
